Include project based on header library dependency
Introduce the ModuleInfo class to contain the
information from the module-info.json file that we need.
In addition to the already-existing project-to-modules
mapping, we add the module-to-project mapping, the
module-to-class mapping, and the module-to-deps mapping.

Extend create_split_manifest to add projects to the
split manifest if there is a dependency from an included
module on a header library module. Use ModuleInfo to
detect this.

Remove manual project additions related to b/151755703.

Bug: 151755703
Test: manifest_split_test
Change-Id: I1f9cfe127e0846ce08055f918425d776ddf605d9