Add build configuration goal selection logic.
This change extends the Treble build sandbox tools to
allow specifying the contexts for each goal. The change
introduces a default built-in context of 'ci'
(continuous integration). For other types of builds
(local, etc.) we can add additional contexts. If a build
configuration goal does not specify contexts, the
selection logic always selects that goal. This behavior
is backwards compatible with any build configurations
that do not contain contexts.

Test: existing test.sh
Test: extended by adding testBuildTargetToBuildGoalsWithContexts
Bug: 155672903
Change-Id: Ib525d9d5e74dbdc24d614fa76823ef55f0b78de5