Generate chd_merged_sepolicy while building CHD
chd_merged_sepolicy is created by merging the system and
vendor sepolicy files of the device build with the vendor
sepolicy files of Cuttlefish.

This approach is a short term solution to the absence of
essential vendor sepolicy rules for device SSI. In the
long term, however, all device sepolicy rules should be
placed at the appropriate partition.

Bug: 315474132
Test: build_cf_hybrid_device
Change-Id: I3c90147ec5ecb326e8c742d4ab41bb16fc3d4382